pyproject.toml
[project]
name = "ai-model-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "dashscope>=1.24.2",
    "openai>=1.101.0",
    "python-dotenv>=1.1.1",
    "qianfan>=0.4.12.3",
    "requests>=2.32.5",
    "zhipuai>=2.1.5.20250825",
]
